+++ CHANGELOG.md
@@ -1,13 +1,80 @@
 # Change Log 
 ## [Unreleased]
 
+## [2.0 beta] - 2021-02-19
+## Added
+- [`README.md`] line 20:
+  >` - [v2.0 beta+][`seqToCanvas_arc.js`](seqToCanvas_arc.js)`
+
+## Removed
+- [`index_main_arc.js`]
+- [`index_arc.html`]
+
+## Changed
+- [`index_main.js`] lines 43-63:  
+  Changes to use 2 or more codes to draw  
+  >` 	//`  
+  >` 	//changing output type`  
+  >` 	slf.getElementById('changeB').addEventListener('click',()=>{`  
+  >` 		//`  
+  >` 		//output types`  
+  >` 		let typeList=['Line','Arc'],L=0,idx=(+outputIndex)+1;`  
+  >` 		//`  
+  >` 		L=typeList.length;`  
+  >` 		//`  
+  >` 		idx=idx<L?idx:0;`  
+  >` 		outputIndex=idx;`  
+  >` 		outputType.textContent=typeList[idx];`  
+  >` 		//`  
+  >` 		text=slf.getElementById('textInput').value;`  
+  >` 		text=!text?'\x00':text;`  
+  >` 		//`  
+  >` 		C.width=slf.getElementById('cWidth').value;`  
+  >` 		C.height=slf.getElementById('cHeight').value;`  
+  >` 		//`  
+  >` 		OUTPUT(text);`  
+  >` 	},false);`
+
+- [`index_main.js`] line 34:  
+  Changes to use 2 or more codes to draw  
+  >` 		OUTPUT(text);`
+
+- [`index_main.js`] lines 11-23:  
+  Changes to use 2 or more codes to draw  
+  >` 	let slf=window.document,cId='outputCvs',C=slf.getElementById(cId),`  
+  >` 		SS=capturingCanvas(C,slf.getElementById('downloadDiv')),text='',`  
+  >` 		outputType=slf.getElementById('seqToCanvas_type'),outputIndex=0,`  
+  >` 		OUTPUT=(txt)=>{`  
+  >` 			//=== output type: 'Line', output index 0 ===`  
+  >` 			if(!(outputIndex!=0)){`  
+  >` 				seqToCanvas.fromText(cId,txt);`  
+  >` 			}`  
+  >` 			//=== output type: 'Arc', output index 1 ===`  
+  >` 			if(!(outputIndex!=1)){`  
+  >` 				seqToCanvas_arc.fromText(cId,txt);`  
+  >` 			}`  
+  >` 		};`
+
+- [`index.html`] lines 57-58:  
+  Changes to use 2 or more codes to draw  
+  >` 		<!-- button to change output type -->`  
+  >` 		<button id='changeB'>Change output type</button>`
+
+- [`index.html`] line 29:  
+  Changes to use 2 or more codes to draw  
+  >` 		<h1>seqToCanvas [type:<code id='seqToCanvas_type'>Line</code>]</h1>`
+
+## Added
+- [`index.html`] line 26:
+  >` 		<script type='text/javascript' src='seqToCanvas_arc.js'></script>`
+
 ## [2.0 beta] - 2021-02-18
-# Fixed
+## Fixed
 - [`algorithm.md`] lines 28-29:
   >`     If it satisfies condition `f0<f||!(f0%(f+1)!=0)`, it closes the current path. `  
   >`     `f` is the current method which is called after methhod `f0`.  `
 
-# Added
+## Added
 - [`algorithm.md`] lines 14-31: description of `seqToCanvas_arc.js`
 - [`index_main_arc.js`]: html interface script for `seqToCanvas_arc.js`
 - [`index_arc.html`]: html interface for `seqToCanvas_arc.js`
@@ -19,7 +86,7 @@
 - [`README.md`] line 14:
   >` ### [Algorithm](algorithm.md)`
 
-# Added
+## Added
 - [`algorithm.md`]: description of algorithm
 
 ## [1.01] - 2021-02-16